Add CONTRIBUTING.md doc
This is essentially to make the comment over current_kernel_abi() more
visible and add some handholding for new developers.  Coming in and
running this crate on an older kernel results in multiple test failures,
and finding the current_kernel_abi() comment takes a little
investigation, so spelling this all out for new contributors may help
save some time.

Signed-off-by: Daniel Burgener <Daniel.Burgener@microsoft.com>